% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{create_tables}
\alias{create_tables}
\title{Load Derived Lego Datasets to Environment}
\usage{
create_tables(sample_data = TRUE)
}
\description{
Loads 'set_words' and the document term matrix 'dtm'.
The 'set_colors' data needs to be in the golbal environment
}
