% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-utils.R
\name{generate_topic_palette}
\alias{generate_topic_palette}
\title{Generate a color for each topic}
\usage{
generate_topic_palette(lda_model, set_colors)
}
\arguments{
\item{set_colors}{The set_colors data frame generated by 'load_data'}

\item{lda_models}{A list of models of class LDA}

\item{model_num}{The mode}
}
\description{
Generate a function that returns the topic palette 
with a color associated with each topic. Function assumes
that it is called within legolda directory.
}
