% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-scores.R
\name{get_lda_clusters}
\alias{get_lda_clusters}
\title{Cluster LDA Models}
\usage{
get_lda_clusters(lda_models)
}
\arguments{
\item{A}{list of LDA models}
}
\value{
A tibble with k, score and metric used
}
\description{
Cluster LDA Models
}
