% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-scores.R
\name{get_topic_numbers}
\alias{get_topic_numbers}
\title{Retrieve Topic Numbers}
\usage{
get_topic_numbers(models)
}
\arguments{
\item{models}{A list of \code{topicmodels::LDA} models}
}
\value{
A list integers
}
\description{
Returns a list of topic numbers from LDA models
}
