% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold.R
\name{kfold}
\alias{kfold}
\title{Create k-fold train and test sets}
\usage{
kfold(data, k = 5, id = ".id")
}
\arguments{
\item{data}{A matrix or dataframe}

\item{k}{An integer: The number of folds.}

\item{id}{}
}
\value{
A list of tibbles with train and test columns.
}
\description{
Based on the function from the modelr package
}
