% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_color_tables}
\alias{load_color_tables}
\title{Load several data sets into the global environment}
\usage{
load_color_tables(sample_data = TRUE)
}
\description{
Loads 'set_words' and the document term matrix 'dtm'.
The 'set_colors' data needs to be in teh golbal environment
}
