% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{load_data}
\alias{load_data}
\title{Load set colors and other data sets from an DB}
\usage{
load_data(sample_data = TRUE, n_samples = 10000)
}
\arguments{
\item{sample_data}{Logical. Only return a sample.}

\item{n_samples}{An integer. The number of samples to return.}
}
\value{
set_colors is assigned the returned data set
}
\description{
Function relies on a connection being returned by the 
 connect function.
}
