% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldatuning2.R
\name{plot_lda_scores}
\alias{plot_lda_scores}
\title{FindTopicsNumber_plot}
\usage{
plot_lda_scores(values, pal = NULL)
}
\arguments{
\item{values}{Data-frame with first column named `topics` and other columns
are values of metrics.}
}
\description{
Support function to analyze optimal topic number. Use output of the
score_models function. I have changed the original
}
\examples{
\dontrun{

library(topicmodels)
data("AssociatedPress", package="topicmodels")
dtm <- AssociatedPress[1:10, ]
optimal.topics <- FindTopicsNumber(dtm, topics = 2:10,
  metrics = c("Arun2010", "CaoJuan2009", "Griffiths2004")
)
FindTopicsNumber_plot(optimal.topics)
}

}
