% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-scores.R
\name{score_lda}
\alias{score_lda}
\title{Score LDA Clusters}
\usage{
score_lda(lda_clusters, total_words)
}
\arguments{
\item{lda_clusters}{: A tibble generated by \code{get_lda_clusters}}

\item{total_words}{: A dataframe generated by \code{load_color_tables}}
}
\value{
return
}
\description{
Scores LDA clusters based on \code{clues::adjustedRand}
}
