% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldatuning2.R
\name{score_models}
\alias{score_models}
\title{Score Models Using Scores from \code{ldatuning} Package}
\usage{
score_models(models, dtm, topics = seq(10, 40, by = 10),
  metrics = c("CaoJuan2009", "Arun2010", "Deveaud2014"), verbose = FALSE)
}
\arguments{
\item{dtm}{An object of class "DocumentTermMatrix" with
term-frequency weighting or an object coercible to a
"simple_triplet_matrix" with integer entries.}

\item{topics}{Vvector with number of topics to compare different models.}

\item{metrics}{String or vector of possible metrics: "Griffiths2004",
"CaoJuan2009", "Arun2010", "Deveaud2014".}

\item{verbose}{Logical; Print status updates.}

\item{mc.cores}{Integer; The number of CPU cores to processes models
simultaneously.}
}
\value{
Data-frame with one or more metrics.  numbers of topics and
  corresponding values of metric. Can be directly used by
  FindTopicsNumber_plot to draw a plot.
}
\description{
Replaces FindTopicsNumber from ldatuning package. Takes
topicmodels LDA output models and topic count and returns
several scores of the clustering fit.
}
