% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-names.R
\name{str_collapse}
\alias{str_collapse}
\title{Collapse String List}
\usage{
str_collapse(str_list, len, max_n = 2)
}
\arguments{
\item{len}{An integer. Length over which to use only the first max_n strings.}

\item{max_n}{An integer. Maximum number of strings to collapse.}
}
\value{
A string which is a comma separated list of collapsed strings.
}
\description{
Collapses a list of strings but limits the number of strings 
 collapsed to max_n
}
