% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-data.R
\name{tidy_colorsets}
\alias{tidy_colorsets}
\title{Load several data sets into the global environment}
\usage{
tidy_colorsets()
}
\description{
Loads 'set_words' and the document term matrix 'dtm'.
The 'set_colors' data needs to be in teh golbal environment
}
