% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-relevance.R
\name{top_terms}
\alias{top_terms}
\title{Compute the Top Terms of an LDA Model}
\usage{
top_terms(model, lambda, nterms, freq)
}
\arguments{
\item{model}{An LDA model produced by the topicmodels package.}

\item{lambda}{A float. Percent to weight each term.}

\item{nterms}{An integer. The number of top terms to return.}

\item{freq}{A table of word per document frequency counts. (TODO)}
}
\value{
A tibble of the top 'nterms' terms for each topic.
}
\description{
Computes a relevance score using the weighted sum 
of the document term score and corpus term freqeuncy.
}
