% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_utils.R
\name{vec_2_function}
\alias{vec_2_function}
\title{Generate a pallete from a color vector}
\usage{
vec_2_function(vec)
}
\arguments{
\item{vec}{A vector to convert ot a function}

\item{fname}{A string. The name of the function
that is created}
}
\value{
Writes a file to a local R directory. Assumes 
 that the function is called from a package directory.
}
\description{
Generate a pallete from a color vector
}
