% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle2.R
\name{waff}
\alias{waff}
\title{A wrapper of the waffle function}
\usage{
waff(data, nrows = 5, size = 0.5, nchr = 20, pad = 0)
}
\arguments{
\item{data}{The data}

\item{nrows}{An integer. The umber of rows of the waffle plot}

\item{size}{A double. Controls the border size of individual squares.}

\item{nchr}{An integer. The number of characters in the title.}

\item{pad}{A double. Plot margin padding.}
}
\value{
return The plot
}
\description{
A wrapper of the waffle function
}
