% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted-relevance.R
\name{weight}
\alias{weight}
\title{Compute Weighted Sum of Vectors}
\usage{
weight(x, y, wt)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A vector
}
\description{
Compute Weighted Sum of Vectors
}
