% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-names.R
\name{get_topic_names}
\alias{get_topic_names}
\title{Get topic names based on an LDA model}
\usage{
get_topic_names(model)
}
\arguments{
\item{model}{The LDA model which will have names created for its 
topics}
}
\value{
A tibble with the single column `topic_name`
}
\description{
Get topic names based on an LDA model
}
