% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle2.R
\name{waff}
\alias{waff}
\title{A wrapper of the waffle function}
\usage{
waff(data, nchr, bgcol, rows, size = 0.2)
}
\arguments{
\item{data}{A tibble with columns, name,theme, year, and counts. 
The counts column is a column of tables.}

\item{nchr}{An integer. The number of characters in the title.}

\item{bgcol}{A valid}

\item{size}{A double. Controls the border size of individual squares.}

\item{row}{An integer. The number of rows of the waffle plot}
}
\value{
return The plot
}
\description{
A wrapper of the waffle function
}
