% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-models.R
\name{weight}
\alias{weight}
\title{Compute weighte sum of vectors}
\usage{
weight(x, y, wt)
}
\arguments{
\item{x}{A vector}

\item{y}{A vector}
}
\value{
A vector
}
\description{
Compute weighte sum of vectors
}
